refactor: cache the dataset
